+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add CORS support
+### Fixed
+- Fix PollResult voted bug
 
 ## [1.8.2] - 2024-07-15
 ### Fixed