+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Consolidate the information, and make it accessible with a single API call [#70](https://github.com/rokwire/polls-building-block/issues/70)
 
 ## [1.8.2] - 2024-07-15
 ### Fixed