+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Consolidate the information, and make it accessible with a single API call [#70](https://github.com/rokwire/polls-building-block/issues/70)
+- Implement /polls API with HTTP POST method [#68](https://github.com/rokwire/polls-building-block/issues/68)
 
 ## [1.8.2] - 2024-07-15
 ### Fixed